question-answering-app/question-answering-data-loader/requirements.txt
gspread==5.12.0
openai==0.28.1
oauth2client==4.1.3
pinecone-client==2.2.4
